git-svn-id: http://svn.vuze.com/public/client@40197 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/com/aelitis/azureus/core/dht/db/DHTDB.java
@@ -57,7 +57,7 @@ DHTDB
 	store(
 		HashWrapper		key,
 		byte[]			value,
-		byte			flags,
+		short			flags,
 		byte			life_hours,
 		byte			replication_control );
 	
@@ -111,7 +111,7 @@ DHTDB
 		DHTTransportContact		reader,
 		HashWrapper				key,
 		int						max_values,
-		byte					flags,
+		short					flags,
 		boolean					external_request );
 		
 		/**
