+++ trunk/azureus2/src/com/aelitis/azureus/core/dht/transport/DHTTransportRequestHandler.java
@@ -68,7 +68,7 @@ DHTTransportRequestHandler
 		DHTTransportContact contact, 
 		byte[]				key,
 		int					max_values,
-		byte				flags );
+		short				flags );
 
 		/**
 		 * Mechanism for reporting that a contact has been imported
